+++ src/components/button/types.ts
@@ -15,4 +15,6 @@ export interface ButtonProps {
     variant?: ButtonVariants
     rounded?: boolean
     onPress?: () => void
+    loading?: boolean
+    loadingText?: string
 }